Updated with inline instructions
For this template to work, you must use this master template to build your stack, but first you need to upload all the other custom templates to your AWS accounts S3 and modify this template to refer to templates 02, 03 and 04 in your S4. There are four references you need to update, one for 02, one for 03 and two for 04.